==> 02-Manipulation-base.Rmd <==

# Manipulation de la base {#c02-Manipulation-base}
<div align="justify">
On peut maintenant regarder à quoi ressemble la table principale, celle des données.
```{r}
RP %>% head() %>% gt()
```

La fonction `str()` ou son équivalent en langage `dplyr` la fonction `glimpse` nous donne la liste des variables avec cette fois leur format et leurs modalités (les premières seulement s'il y en a beaucoup).
```{r}
# str(RP)
RP %>% glimpse()
```
Les variables sont en format, soit 'int' ("integer" -nombres entiers), soit 'fct' ("factor"), ou encore 'dbl' ("double" - nombres réels).  
On remarque que la plupart des variables ont des modalités en chiffres et/ou lettres, ce qui signifie qu'il faudra probablement recoder ces modalités pour leur donner des libellés dans nos analyses descriptives (tableaux ou graphiques).  

Par ailleurs, certaines variables ont été laissées en format 'integer' bien que l'on avait mis l'option `stringsAsFactors=TRUE` lors du chargement des données, or ces chiffres correspondent bien à des modalités d'une variable catégorielle, il faudra donc le corriger ultérieurement.  

Dernière remarque, si on avait eu des valeurs manquantes sur certaines variables, elles se seraient ici affichées avec la modalité `NA`, cela ne semble pas le cas, à vérifier par la suite.  
  
  
## Format des données - WIDE ou LONG
La question du **format de la table de données** est importante, on distingue généralement le format *wide* du format *long* : 

- le premier renvoie au cas où chaque ligne correspond a un et un seul individu (ou autre niveau de l'analyse : une entreprise, un ménage, un logement, etc.), et toutes les informations le concernant sont données en colonnes (ce qui peut augmenter rapidement le nombre de colonnes), c'est l'exemple d'une information que l'on a sur plusieurs années, chaque colonne correspondant à l'information de l'année en question ; 
- le second, au contraire, multiplie les lignes pour un seul individu, mais il y a alors moins de colonnes, à partir du même exemple précédent, il y aura cette fois une colonne indiquant l'année et une colonne contenant l'information.  

Ici, chaque ligne correspond à un **logement** auquel est associé un **ménage** (ou le contraire si vous préférez) ; on est bien dans un format `wide`.   

En revanche, si on considère que notre niveau d'analyse est la commune, on voit alors qu'on dispose de plusieurs lignes par commune et si l'on veut garder quelques informations à ce niveau en n'ayant plus qu'une ligne par commune, on peut changer de format.  

Le package `tidyverse` contient deux fonctions qui permettent de passer d'un format à un autre : `pivot_longer()` et `pivot_wider()` présentés en détails <a href="https://tidyr.tidyverse.org/articles/pivot.html" target="_blank">ici</a>. Nous aurons l'occasion de les utiliser ultérieurement dans ce cours.  
  


## Sélectionner un champ réduit de la base, séparer les données
Manipuler une base de données, c'est aussi travailler ou créer une table réduite, surtout quand on a des données volumineuses comme ici : c'est donc choisir un champ particulier de la base.  

On peut par exemple ici choisir un champ géographique plus réduit au sein de l'Île-de-France, à partir de la commune.   

On peut aussi choisir de séparer les données en créant 2 tables selon l'information donnée : le logement ou la personne de référence du ménage.  

Comme vous le savez très probablement, dans le langage `tidyverse`, pour faire des sélections sur les lignes ou observations dans une table de données, on utilise la fonction `filter()` ; pour faire des sélections sur les colonnes ou variables, on utilise la fonction `select()`  ; pour supprimer des variables, on utilise également `select()` avec un "-" devant la (ou les) variable(s) à supprimer (dans le langage `R` de base, on peut aussi utiliser la fonction `subset()`).   
   

**EXERCICES** :  
Créer trois nouvelles bases selon les indications suivantes :

- la première que vous nommerez "RP_Paris" et qui rassemblera les données pour le seul département de Paris ;
- la deuxième que nous nommerez "RP_menage" et qui comprendra les données pour les seules variables relatives à la personne de référence du ménage ; 
- la troisième que vous nommerez "RP_log_Paris" et qui rassemblera les données pour le seul département de Paris et comprendra toutes les variables sauf celles relatives à la personne de référence du ménage.

<!-- Les premières lignes de ces trois tables devront ressembler à cela : --> 
```{r echo=FALSE, include=FALSE}
RP_Paris <- RP %>% filter(COMMUNE == "75056")

RP_menage <- RP %>% select("AEMM", "AEMMR", "AGEMEN8", "ANEM", "ANEMR", "DIPLM", 
                           "EMPLM", "ILETUDM", "ILTM", "IMMIM", "INAIM", "IRANM",
                           "RECHM", "SEXEM", "STAT_CONJM", "TACTM", "TPM", "TRANSM")
  
RP_log_Paris  <- RP %>% filter(COMMUNE == "75056") %>% 
                        select(-c("AEMM", "AEMMR", "AGEMEN8", "ANEM", "ANEMR", "DIPLM", 
                           "EMPLM", "ILETUDM", "ILTM", "IMMIM", "INAIM", "IRANM",
                           "RECHM", "SEXEM", "STAT_CONJM", "TACTM", "TPM", "TRANSM"))
RP_Paris %>% head(4) %>% gt()
RP_menage %>% head(4) %>% gt()
RP_log_Paris%>% head(4) %>% gt()
```


<div class="solution-exo">
```{r eval=FALSE}
RP_Paris <- RP %>% filter(COMMUNE == "75056")

RP_menage <- RP %>% select("AEMM", "AEMMR", "AGEMEN8", "ANEM", "ANEMR", "DIPLM", 
                           "EMPLM", "ILETUDM", "ILTM", "IMMIM", "INAIM", "IRANM",
                           "RECHM", "SEXEM", "STAT_CONJM", "TACTM", "TPM", "TRANSM")
  
RP_log_Paris <- RP %>% filter(COMMUNE == "75056") %>% 
                        select(-c("AEMM", "AEMMR", "AGEMEN8", "ANEM", "ANEMR", "DIPLM", 
                           "EMPLM", "ILETUDM", "ILTM", "IMMIM", "INAIM", "IRANM",
                           "RECHM", "SEXEM", "STAT_CONJM", "TACTM", "TPM", "TRANSM"))
```
</div>
\fi


```{r eval=FALSE, include=FALSE}
RP_Paris <- subset(RP, COMMUNE == "75056")
RP_menage <- subset(RP, select=c("AEMM", "AEMMR", "AGEMEN8", "ANEM", "ANEMR", "DIPLM", 
                           "EMPLM", "ILETUDM", "ILTM", "IMMIM", "INAIM", "IRANM",
                          "RECHM", "SEXEM", "STAT_CONJM", "TACTM", "TPM", "TRANSM"))
```




<!-- This line is commented, won't show in the document -->

```{r eval=FALSE, include=FALSE}
str(RP$COMMUNE)
RP <- RP %>% mutate(COMMUNE=as.factor(COMMUNE))
levels(RP$COMMUNE)
```

```{r eval=FALSE, include=FALSE}
# Exemple agrégation données sur commune et utilisation de la fonction ``pivot_wider`
RP %>% select(COMMUNE, TYPL, IPONDL) %>%
  group_by(COMMUNE) %>% 
  count(TYPL, wt=IPONDL) %>%
  pivot_wider(names_from = TYPL, values_from = n, names_prefix = "TYPL_")
```
